Fix group sync: use CN as source of comparison with LibreBooking's groups
The former code used to use the full DN (distinguishedName - e.g. "CN=My Group,OU=Groups OU,DC=ldap,DC=mycompany,DC=com) and I instead extracted only the CN (common name - e.g. "My group") so that it can really be compared to LB's group names.

This was an issue with Active Directory (for which the current Plugin didn't work for me).
Not sure whether this needs to be fixed for other LDAP implementation.